# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Onyx Domain API V1",
    "version" : "1.0.9"
  },
  "servers" : [ {
    "url" : "http://onyxdomainapi.azurewebsites.net"
  } ],
  "paths" : {
...
    "/Loans/List" : {
      "get" : {
        "tags" : [ "Loans" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Loans/Detail" : {
      "get" : {
        "tags" : [ "Loans" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Loans/Delete" : {
      "post" : {
        "tags" : [ "Loans" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Loans/Save" : {
      "post" : {
        "tags" : [ "Loans" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoanData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoanData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoanData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
    "/Members/Save" : {
      "post" : {
        "tags" : [ "Members" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Members/Delete" : {
      "post" : {
        "tags" : [ "Members" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/ModelUploads/List" : {
      "get" : {
        "tags" : [ "ModelUploads" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/ModelUploads/Download" : {
      "get" : {
        "tags" : [ "ModelUploads" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/octet-stream" : { }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Context" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "nullable" : true
          },
          "success" : {
            "type" : "boolean"
          },
          "data" : {
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "data" : "",
          "success" : true,
          "message" : "message"
        }
      },
...
      "IdOnly" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
...
      "LoanData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "rate" : {
            "type" : "number",
            "format" : "double",
            "nullable" : true
          },
          "tradeDate" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "expiryDate" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "type" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "brokerId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "ticker" : {
            "type" : "string",
            "nullable" : true
          },
          "rolloverNumber" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "MemberData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "code" : {
            "type" : "string",
            "nullable" : true
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "email" : {
            "type" : "string",
            "nullable" : true
          },
          "analyst" : {
            "type" : "boolean"
          },
          "trader" : {
            "type" : "boolean"
          },
          "disabled" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}